+++ packages/tinacms/src/admin/pages/LogoutPage.tsx
@@ -12,19 +12,19 @@ limitations under the License.
 */
 
 import React from 'react'
-import { BiLogIn } from 'react-icons/bi'
+import { BiLogOut } from 'react-icons/bi'
 import { MdOutlineArrowBack } from 'react-icons/md'
 import { setEditing } from '../../edit-state'
 import AuthTemplate from '../components/AuthTemplate'
 
-const login = () => {
-  setEditing(true)
-  window.location.reload()
+const logout = () => {
+  setEditing(false)
+  window.location.href = '/'
 }
 
-const LoginPage = () => {
+const LogoutPage = () => {
   return (
-    <AuthTemplate message="Please log in to Tina Cloud to access your content.">
+    <AuthTemplate>
       <a
         href="/"
         className="flex-1 text-center inline-flex justify-center items-center px-8 py-3 shadow-sm text-sm leading-4 font-medium rounded-full text-gray-600 border border-gray-150 hover:opacity-80 hover:bg-gray-50 focus:outline-none focus:shadow-outline-blue  transition duration-150 ease-out"
@@ -34,14 +34,14 @@ const LoginPage = () => {
       </a>
       <button
         type="submit"
-        onClick={() => login()}
+        onClick={() => logout()}
         className="flex-1 justify-center text-center inline-flex items-center px-8 py-3 shadow-sm border border-transparent text-sm leading-4 font-medium rounded-full text-white hover:opacity-80 focus:outline-none focus:shadow-outline-blue  transition duration-150 ease-out"
         style={{ background: '#0084FF' }}
       >
-        <BiLogIn className="w-6 h-auto mr-1.5 opacity-80" /> Log in
+        <BiLogOut className="w-6 h-auto mr-1.5 opacity-80" /> Log out
       </button>
     </AuthTemplate>
   )
 }
 
-export default LoginPage
+export default LogoutPage
